Moving tests to storage.rs as unit tests gives access to what we need
