+++ storage/tests/test_mvcc.rs
@@ -36,101 +36,4 @@ This file should comprise a set of low-level tests relating to MVCC.
 
  */
 
-use std::path::PathBuf;
-use std::rc::Rc;
-
-use bytes::byte_array::ByteArray;
-use bytes::byte_reference::ByteReference;
-use storage::key_value::{StorageKey, StorageKeyArray, StorageKeyReference};
-use storage::keyspace::keyspace::KeyspaceId;
-use storage::MVCCStorage;
-use test_utils::{create_tmp_dir, init_logging};
-
-const KEYSPACE_ID: KeyspaceId = 0;
-const KEY_1: [u8; 4] = [0x0, 0x0, 0x0, 0x1];
-const KEY_2: [u8; 4] = [0x0, 0x0, 0x0, 0x2];
-const VALUE_0: [u8; 1] = [0x0];
-const VALUE_1: [u8; 1] = [0x1];
-const VALUE_2: [u8; 1] = [0x1];
-fn setup_storage(storage_path: &PathBuf) -> MVCCStorage {
-    let mut storage = MVCCStorage::new(Rc::from("storage"), &storage_path).unwrap();
-    storage.create_keyspace("keyspace", KEYSPACE_ID, &MVCCStorage::new_db_options()).unwrap();
-    storage
-}
-
-#[test]
-fn test_commit_increments_watermark() {
-    init_logging();
-    let storage_path = create_tmp_dir();
-    let storage = setup_storage(&storage_path);
-    let snapshot_0 = storage.open_snapshot_write();
-    // TODO: How does one access the watermark?
-    snapshot_0.commit();
-}
-
-#[test]
-fn test_reading_snapshots() {
-    init_logging();
-    let storage_path = create_tmp_dir();
-    let storage = setup_storage(&storage_path);
-
-    let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(KEYSPACE_ID, ByteReference::new(&KEY_1)));
-
-    let snapshot_write_0 = storage.open_snapshot_write();
-    snapshot_write_0.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_0));
-    snapshot_write_0.commit();
-
-    let snapshot_read_0 = storage.open_snapshot_read();
-    assert_eq!(snapshot_read_0.get::<128>(key_1).unwrap().bytes(), VALUE_0);
-
-    let snapshot_write_1 = storage.open_snapshot_write();
-    snapshot_write_1.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_1));
-    let snapshot_read_01 = storage.open_snapshot_read();
-    assert_eq!(snapshot_read_0.get::<128>(key_1).unwrap().bytes(), VALUE_0);
-    assert_eq!(snapshot_read_01.get::<128>(key_1).unwrap().bytes(), VALUE_0);
-
-    let result_write_1 = snapshot_write_1.commit();
-    assert!(result_write_1.is_ok());
-
-    let snapshot_read_1 = storage.open_snapshot_read();
-    assert_eq!(snapshot_read_0.get::<128>(key_1).unwrap().bytes(), VALUE_0);
-    assert_eq!(snapshot_read_01.get::<128>(key_1).unwrap().bytes(), VALUE_0);
-    assert_eq!(snapshot_read_1.get::<128>(key_1).unwrap().bytes(), VALUE_1);
-    snapshot_read_1.close();
-    snapshot_read_01.close();
-    snapshot_read_0.close();
-}
-
-
-#[test]
-fn test_write_conflicts() {
-    init_logging();
-    let storage_path = create_tmp_dir();
-    let storage = setup_storage(&storage_path);
-
-    let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(KEYSPACE_ID, ByteReference::new(&KEY_1)));
-    let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(KEYSPACE_ID, ByteReference::new(&KEY_2)));
-
-    let snapshot_write_0 = storage.open_snapshot_write();
-    snapshot_write_0.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_0));
-    snapshot_write_0.commit();
-
-    let snapshot_write_11 = storage.open_snapshot_write();
-    let snapshot_write_21 = storage.open_snapshot_write();
-    snapshot_write_11.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_1));
-    snapshot_write_21.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_2));
-    let result_write_11 = snapshot_write_11.commit();
-    assert!(result_write_11.is_ok());
-    let result_write_21 = snapshot_write_21.commit();
-    // assert!(!result_write_21.is_ok()); // Fails
-
-    let snapshot_write_12 = storage.open_snapshot_write();
-    let snapshot_write_22 = storage.open_snapshot_write();
-    snapshot_write_12.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_2)), ByteArray::copy(&VALUE_1));
-    snapshot_write_22.put_val(StorageKeyArray::new(KEYSPACE_ID, ByteArray::copy(&KEY_2)), ByteArray::copy(&VALUE_2));
-    let result_write_22 = snapshot_write_22.commit();
-    assert!(result_write_22.is_ok());
-    let result_write_12 = snapshot_write_12.commit();
-
-    // assert!(!result_write_12.is_ok()); // Fail
-}
+// Moved to storage.rs
